Handle larger rootfs
As we added funcitonality to run our proc program inside buildroot
we bloated the size of our rootfs from 750KB to 35MB.

The initram's location in in guest PA is fixed and we are not sure
what needs to change to fix this.
Since the initramfs(i.e. rootfs) grew it overapped with DTB location.

This needed fixing in many places:
- Move DTB to a location before initrams
- change guest RAM in our VMM Code as well as the linux.dts
- account for a larger initramfs file in the linux.dts