Enable access to /dev/mem in qemu image
Writing to /dev/mem directly on Linux is not allowed by default.
However, it cannot be enabled with sudo or sysctl. You need to recompile the kernel.
See the change in the linux_config file to see what needs to change for
recompilation

But that's not all.
Enabling writing to /dev/mem only allowed you to write to it as a file, i.e., write syscalls.
But if you want to write to it by mapping /dev/mem using mmap,
you need to boot the kernel with "nopat" boot option. This option disables Page Attributes.
This took a few more hours of Googling.

Done only for qemu for now, and not odroid.